refactor: #defines in Heuristics and new file Constants.cpp
bench 2855713
